Вынести логику получения статуса задания из ВИС в абстрактный класс
